KKS port and project code merge (#9)
- added Stiletto.Core project
- moved code from Stiletto project to Stiletto.Core updated dependencies
- added configuration for conditional compilation
- added Stiletto.KKS updated dependencies
- added conditional compilation steps to compile the plug-in for KKS
- Did not look in depth into Stiletto_KKS since the code is so different it was easier to work on the Stiletto code base.

I don't use Studio so no testing there. It seems to work fine in KKS.

For the method OnHeelInfoUpdate had to use try {} block. It gives and error the second time you invoke Maker in the same
game session. Don't have KK installed for testing this change in the code for KK. Though logic dictates that the problem should be present in KK also. The nullchecks fix if active will permit the plug-in to continue to run as is the case in KKS.