Shoe warp: fix communication (#10)
Upon entering a Communication scene, the shoe state was getting reset
sufficiently to wipe out the changes we made to the bind poses, but not
sufficiently to cause our dirty flag to get set. The animation state does
get reset in this case, so we can reuse that detection to avoid breakage.